Update dependencies
So the developer of the telegrambots library did some strange stuff with fragmentation of his library on many pieces but he didn't provided enough clarity in documentation. That is why at first I used some outdated libraries. Now it is updated to last version and it works with Eureka fine.